+++ mitmproxy/addonmanager.py
@@ -1,17 +1,20 @@
 import contextlib
 import inspect
+import logging
 import pprint
-import sys
 import traceback
 import types
 from collections.abc import Callable, Sequence
 from dataclasses import dataclass
 from typing import Any, Optional
 
-from mitmproxy import hooks
+import sys
+
 from mitmproxy import exceptions
 from mitmproxy import flow
-from . import ctx
+from mitmproxy import hooks
+
+logger = logging.getLogger(__name__)
 
 
 def _get_name(itm):
@@ -48,7 +51,7 @@ def safecall():
         etype, value, tb = sys.exc_info()
         tb = cut_traceback(tb, "invoke_addon_sync")
         tb = cut_traceback(tb, "invoke_addon")
-        ctx.log.error(
+        logger.error(
             "Addon error: %s" % "".join(traceback.format_exception(etype, value, tb))
         )
 
@@ -88,7 +91,7 @@ class Loader:
             if same_signature:
                 return
             else:
-                ctx.log.warn("Over-riding existing option %s" % name)
+                logger.warning("Over-riding existing option %s" % name)
         self.master.options.add_option(name, typespec, default, help, choices)
 
     def add_command(self, path: str, func: Callable) -> None:
@@ -160,17 +163,18 @@ class AddonManager:
         """
         api_changes = {
             # mitmproxy 6 -> mitmproxy 7
-            "clientconnect": "client_connected",
-            "clientdisconnect": "client_disconnected",
-            "serverconnect": "server_connect and server_connected",
-            "serverdisconnect": "server_disconnected",
+            "clientconnect": f"The clientconnect event has been removed, use client_connected instead",
+            "clientdisconnect": f"The clientdisconnect event has been removed, use client_disconnected instead",
+            "serverconnect": "The serverconnect event has been removed, use server_connect and server_connected instead",
+            "serverdisconnect": f"The serverdisconnect event has been removed, use server_disconnected instead",
+            # mitmproxy 8 -> mitmproxy 9
+            "add_log": "The add_log event has been deprecated, use Python's builtin logging module instead",
         }
         for a in traverse([addon]):
-            for old, new in api_changes.items():
+            for old, msg in api_changes.items():
                 if hasattr(a, old):
-                    ctx.log.warn(
-                        f"The {old} event has been removed, use {new} instead. "
-                        f"For more details, see https://docs.mitmproxy.org/stable/addons-events/."
+                    logger.warning(
+                        f"{msg}. For more details, see https://docs.mitmproxy.org/dev/addons-api-changelog/."
                     )
             name = _get_name(a)
             if name in self.lookup: